back to input stream
